move files | fix user->project relationship
