+++ packages/database/src/config/orm.config.ts
@@ -11,7 +11,6 @@ import { PostgreSqlDriver } from '@mikro-orm/postgresql';
 import { getEnv } from '../env';
 
 export const migrationsPath = `${__dirname}/../migrations`;
-// export const entitiesPath = [`${__dirname}/../entities`,`!${__dirname}/../entities/utils`];
 export const entitiesPath = [`${__dirname}/../entities`];
 
 export const baseExtensions = [Migrator];
