+++ packages/database/src/entitiesUtils/scoreVotes.ts
